requirements.txt
streamlit==1.28.0
pandas==2.0.3
openpyxl==3.1.2
xlsxwriter==3.1.2
gspread==5.11.0
google-auth==2.23.0
python-dateutil==2.8.2